Added some guesswork to opening the frame buffer device.
Open /dev/fbX iteratively instead of failing if /dev/fb is not found
and the device is not configured in xorg.conf. Maintain backward
compatibility by first trying the plain /dev/fb.

Store the found device path and use it in OMAPFBPreInit instead of
trying to guess it again there.